refactor: :recycle: centralize validation functions
