+++ server/utils/zenodo.ts
@@ -1,4 +1,5 @@
 import { buildPosterJson } from "./buildPosterJson";
+import { isValidOrcidChecksum, validateOrcidExists } from "#shared/utils/orcid";
 import isoLanguages from "../../app/assets/data/iso-639-1.json";
 
 const ISO639_1_TO_3: Record<string, string> = Object.fromEntries(
@@ -349,6 +350,28 @@ export async function beginZenodoPublication(
     : [];
 
   const creators = meta.creators as InvenioCreator[];
+
+  // Check ORCID IDs exist in the registry. Fail-open: warns but never blocks publication.
+  const orcidIds = (creators ?? [])
+    .flatMap((c) => c.nameIdentifiers?.map(extractOrcid).filter(Boolean) ?? [])
+    .filter((id): id is string => !!id);
+
+  if (orcidIds.length > 0) {
+    const results = await Promise.all(
+      orcidIds.map((id) => validateOrcidExists(id)),
+    );
+    const invalid = orcidIds.filter((_, i) => !results[i]);
+    if (invalid.length > 0) {
+      console.warn(
+        `[Zenodo] ORCID registry check failed for: ${invalid.join(", ")}`,
+      );
+      await onProgress?.({
+        step: "upload_metadata",
+        status: "error",
+        message: `Could not verify the following ORCID IDs in the registry: ${invalid.join(", ")}. Publication will continue — please confirm these IDs are correct.`,
+      });
+    }
+  }
 
   const rawRelated = meta.relatedIdentifiers as {
     relatedIdentifier?: string;
@@ -902,19 +925,6 @@ type InvenioCreator = {
   }[];
 };
 
-// ORCID format: 4 groups of 4 digits, last char may be X
-const ORCID_PATTERN = /^\d{4}-\d{4}-\d{4}-\d{3}[\dX]$/;
-
-// ISO 7064 mod 11,2 checksum used by ORCID
-function validateOrcidChecksum(bare: string): boolean {
-  const digits = bare.replace(/-/g, "");
-  let total = 0;
-  for (let i = 0; i < 15; i++) total = (total + parseInt(digits[i]!, 10)) * 2;
-  const expected = (12 - (total % 11)) % 11;
-
-  return digits[15] === (expected === 10 ? "X" : String(expected));
-}
-
 // Build InvenioRDM-format creator objects from DB creator data.
 // DB data has full structured creators; legacy draft only has combined name strings.
 function buildRdmCreators(
@@ -929,11 +939,7 @@ function buildRdmCreators(
     const orcidRaw = c.nameIdentifiers?.map(extractOrcid).find(Boolean);
     const orcidBare = orcidRaw?.replace(/^https?:\/\/orcid\.org\//i, "").trim();
     const orcid =
-      orcidBare &&
-      ORCID_PATTERN.test(orcidBare) &&
-      validateOrcidChecksum(orcidBare)
-        ? orcidBare
-        : undefined;
+      orcidBare && isValidOrcidChecksum(orcidBare) ? orcidBare : undefined;
 
     if (orcidRaw && !orcid) {
       console.warn(
